+++ exporter/otlphttpexporter/factory.go
@@ -22,6 +22,7 @@ import (
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
+	"go.opentelemetry.io/collector/config/configcompression"
 	"go.opentelemetry.io/collector/config/confighttp"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
@@ -51,6 +52,8 @@ func createDefaultConfig() config.Exporter {
 			Endpoint: "",
 			Timeout:  30 * time.Second,
 			Headers:  map[string]string{},
+			// Default to gzip compression
+			Compression: configcompression.Gzip,
 			// We almost read 0 bytes, so no need to tune ReadBufferSize.
 			WriteBufferSize: 512 * 1024,
 		},